backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
redis==5.2.0
celery==5.4.0
pydantic-settings==2.6.1
alembic==1.14.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.27.2
cryptography==44.0.2
croniter==3.0.3
jsonschema==4.23.0
prometheus-client==0.21.1
